% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bpaearmtype}
\alias{bpaearmtype}
\title{bpaearmtype}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 938 rows and 10 columns.}
\usage{
bpaearmtype
}
\description{
bpaearmtype
}
\keyword{datasets}
