% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bpaelong}
\alias{bpaelong}
\title{bpaelong}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 45725 rows and 6 columns.}
\usage{
bpaelong
}
\description{
bpaelong
}
\keyword{datasets}
