% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bpaeraw}
\alias{bpaeraw}
\title{bpaeraw}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 59 rows and 1025 columns.}
\usage{
bpaeraw
}
\description{
bpaeraw
}
\keyword{datasets}
