% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bpaeriskdiff}
\alias{bpaeriskdiff}
\title{bpaeriskdiff}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 352 rows and 11 columns.}
\usage{
bpaeriskdiff
}
\description{
bpaeriskdiff
}
\keyword{datasets}
